Fix disputing blocks with genesis block as parent
The disputeBlockHeader contract function expects the RLP encoded header of the parent of the disputed header.
Since the RLP headers are retrieved by checking the logs for events that are only emitted on submittal of a block after initial contract deployment, the genesis block can't be found by this method.

Therefore an event is now also submitted in the constructor.